tidy: Use the 'override' keyword on destructors.
The new override keyword should be used on functions in a derived
classes that overload a function in their base class.  It implicitly
declares the function as virtual, so it replaces that keyword.  It
additionally requires that a function exist in the base class with the
exact same type signature.  I.E. it validates that the overridden
function exists.  This prevents functions that look like they override
a base class function but don't.

https://clang.llvm.org/extra/clang-tidy/checks/modernize-use-override.html