tidy: Don't need to check pointer validity.
The clang-tidy "clang-analyzer-core.CallAndMessage" check pointed out
that it thought the m_frameTimer->stop function could be called when
m_frameTimer was null.  Looking at the code, m_frameTimer is created
in the parent constructor, has the deleteLater function called on it
in the parent destructor, and no other reference is protected by a
nullptr check.  It seems that the nullptr check is superfluous, not
that the function can be called through a nullptr.