tidy: Fix 'value stored is never read' warning in audiooutput_omx.cpp.
The clang-tidy "dead stores" check pointed out a variable that is set
and never used, because the code block that contains it is
unconditionally exited.  This is only true because a couple of code
fragments aren't being compiled, and those fragments allow the code
block to complete without exiting.  Move the variable assignment
inside the conditionally compiled code to prevent this false positive.