Show delivery system in "Input connections" configuration screen
Previously the mythtv-setup "Input connections" configuration screen showed
on the first line the "Input name" with user selectable value in drop down box.
For DVB cards the label is now "Delivery system" and the presented value is not editable.
Note that the delivery system can be selected in the card configuration screen.
For non-DVB cards nothing has changed.

Refs #13014