tidy: Ignore return value on some ALSA calls to retrieve data.
The clang-tidy "dead stores" check pointed out four calls to retrieve
data from ALSA whose return values are stored, and then overwritten by
subsequent calls without checking.  This is the way the code was
originally written.  These calls are only used to print data to the
log files, so it seems safe to ignore the errors. At worst, a value in
a log message might be wrong.