tidy: Protect object dereference in ~BonjourRegister.
The clang-tidy "call and message" check pointed out the possibility of
trying to call a function through a nullptr.  The BonjourRegister
object contains a socket, and two of the socket functions are called
from this destructor.  One call is protected by a nullptr check, the
other isn't.  Group both of these socket object calls under the same
validity check.