tidy: Stop processing MythSystemLegacyUnix::Fork if malloc fails.
The clang-tidy "non-null param checker" pointed out that this function
can attempt to call execv with a null parameter, even though that
function's arguments are marked as requiring non-null parameters.  Fix
the code so that if it is unable to allocate memory to copy the
command arguments, that it returns immediately instead of continuing.