tidy: Fix simple warnings of 'value stored is never read' in freesurround.
The clang-tidy "dead stores" check pointed out several places where a
variables was set and never used.  It appears that his code used to be
part of a loop where the setting the variable would have made sense,
but now is only a single pass and the new value is never references.
Clean up these warnings by removing the assignments.