tidy: Fix unused variables in jobqueue.cpp.
The clang-tidy "dead stores" checker pointed out that the
jobsAreRunning variable is never checked after being set to false.
That is because the code immediately breaks out of the loop that is
checking this variable.  Replace the break with a continue, as this
has the same effect but reads better than completely removing the
jobsAreRunning variable.