tidy: Fix unused variables in livetvchain.cpp.
The clang-tidy "uninitialized assignment" checker pointed out a
possibility of an uninitialized variable being used in JumpToNext.
The 'm_curpos' variable should always be limited to the range [0,
m_chain.size), so the while loop should always be executed.  Add an
initialization of the 'switchto' variable so that if the 'm_curpos'
variable somehow gets out of its valid range then this function will
still perform deterministic.

Also duplicate an existing test for ProgramIsAt() returning -1, and if
so, resetting 'm_curpos' to zero.