DVB tuner delivery system read from device
The delivery system, which is used to determine
the tuner type, is now read back from the tuner after the
tuner has been configured with the value in the database
table capturecard/inputname.

If the database does not have a valid value then
the tuner is not configured and the tuner type is
determined from the delivery system that is currently
configured in the tuner, either by power-on default
or by the dvb-fe-tool.

Previously, only the delivery system value from the
database was used to determine the tuner type.
This failed if there was no delivery system
configured in the database for that tuner.

Fixes #13434