tidy: Remove unused code from AvFormatDecoderBD::BDFindPosition.
The clang-tidy "dead stores" checker pointed out that a variable in
this function is set, but that value is never used.  Investigating the
code, there is a ten line code block that computes the value of
'desiredTimePos', and then the value is ignored.  The entire ten lines
can be removed.  They are now conditionally compiled out in case
someone wants to revisit whether the function should be returning a
value based on 'desiredTimePos' or 'desiredFrame'.