tidy: Add missing 'else' in TVRec::GetMaxBitrate.
The clang-tidy "dead stores" checker pointed out that a variable in
the TVRec::GetMaxBitrate function is set, but that value is never
used.  Investigating the code, when a recording profile for HDPVR
support was added in 2008 an 'if blah' was added to the code in the
middle of a series of 'else if blah' statements.  Adding the 'else'
missing keyword eliminates the warning message.

This errant code didn't cause a user visible error as the value in
the final 'else' clause matched the value in the earlier dead code.