tidy: Fix dead stores in mythrender_vdpau.cpp.
The clang-tidy "dead stores" checker pointed out four places that
variables in this file are set but never used.  Two can be fixed by
removing the sets.  The other two are fixed by converting them to use
a macro to load function addresses, like all the other address loads
in that function.