tidy: Fix unused variables in myth archive helper.
The clang-tidy "dead stores" checker pointed out three places where a
variable was set but then never used.  This code reads both the name
and id of the genre (for example), and then ignores the id using only
the name to look up information.  Remove the initial settings for
these variables because they are overwritten in all cases.