tidy: Fix memory leak warnings in SignalMonitor::Init.
The clang-tidy "new/delete leak" checker pointed out several possible
object leaks in the SignalMonitor Init code.  There is no leak here.
This is an artifact of the way the code is written, because all of the
conditionally compiled tests have to assume that they might be the
first test.  Starting with a fake case allows all of the conditionally
compiled cases to start with an 'else if' test, which eliminates the
false positive warnings about leaked objects.