tidy: Fix memory leak warnings in mythcommflag.
The clang-tidy "new/delete leak" checker pointed out that the
print_comm_flag_output might be attempting to delete the global
variable 'cout'.  The complaint occurs because clang-tidy doesn't seem
to remember previous branch choices.  It takes the false branch for
the first of the two 'if (output_filename != "-")' clauses and the
true branch for the other, even thought none of the intervening code
resets the variable.  Using a slightly different second test achieves
the same ends, without confusing the clang-tidy checker.