Use cardid and inputname (delivery system) on OSD when displayname is empty.
In mythtv-setup, an input connection can be given a "Display name (optional)".
This name is typically used to give a user-friendly name to a tuner card.

This name is shown on the screen the first time Live TV is started.

If there is no displayname defined there is a name generated consisting
of the input number, the card type and the inputname.
The inputname is now used to store the delivery system; this leads
to names as "1: DVB/DVB-T2" which contain one DVB too many.
This fix changes the name to the input number and the inputname,
for instance: "1: DVB-T2".
N.B. This is the same format what would be generated by function
TV::UpdateOSDInput in tvplay.cpp, if CardUtil::GetDisplayName would
return an empty string.

Refs #13014