tidy: Fix memory leak in lcdserver main function.
The clang-tidy "new/delete leak" checker pointed out that the
lcdserver main function creates an LCDServer object but never deletes
it.  This is a trivial problem, as the lcdserver is exiting when it
fails to free the memory.  The memory will be freed when the process
terminates.  Add a proper delete call for this object so that code
analyzers don't complain.