tidy: Fix memory leak in mythtranscode main function.
The clang-tidy "new/delete leak" checker pointed out that the
mythtranscode main function creates a ProgramInfo object, but in an
error condition fails to delete it.  This is a trivial problem, as
mythtransode is exiting when it fails to free the memory.  The memory
will be freed when the process terminates.  Add a proper delete call
for this object so that code analyzers don't complain.