tidy: Annotate code that there is no memory leak.
The clang-tidy "new/delete leak" checker found a couple of places in
mythfrontend where it believes there is an object leak.  These do not
appear to be actual leaks as the object constructor adds itself to a
list maintained by its parent.  Annotate the code with this
information, so that next time leaks are being this information will
be easily available.