tidy: Fix memory leak in mythbacked mediaserver code.
The clang-tidy "new/delete leak" checker pointed out that the
WebSocketServer created in the mediaserver initialization didn't have
a corresponding delete.  This is a trivial problem, as the server
should last until the backend shuts down.  Add code to remember the
pointer to the web socket server on the MediaServer object, and then
delete it from the MediaServer destructor.