tidy: Fix memory leak in mythbacked helpers code.
The clang-tidy "new/delete leak" checker pointed out that the system
event handler created in the run_backend function didn't have a
corresponding delete.  This is a trivial problem, as the object should
last until the backend shuts down.  Create a global pointer to this
object, like all the other objects in this function, and delete it
when the programs is being torn down.