tidy: Fix a bunch of null dereference warnings related to MPEG descriptors.
The clang-tidy "null dereference" checker pointed out a bunch of
potential null dereferences in the mpeg descriptor/tables code.  These
are the result of the code creating various objects, and then blindly
dereferencing the _data pointer in those objects (mostly via member
functions).  The problem with those dereferences is that the object
constructor clears the _data pointer if the information passed into
the constructor was too short.  Add a number of tests to check the
validity of the _data pointer before dereferencing it.

This change also tightens up the code to select Australian and Greek
category names from the EIT information, replacing switch statements
with table lookups.