tidy: Fix return value from DecoderBase SetTrack and AutoSelectTrack.
The clang-tidy "implicit boolean conversion" check pointed out a
couple of places in this file where functions return a boolean that
gets implicitly cast to an integer.  According to the documentation
the return value is supposed to be a track number (or -1).  This bug
appears to have been introduced by work done in 2006, but was never
found because all callers of these function ignore the returned
values.

https://clang.llvm.org/extra/clang-tidy/checks/readability-implicit-bool-conversion.html