tidy: Change spinbox setting calls to use 0/1 instead of false/true.
The clang-tidy "implicit boolean conversion" check pointed out a
couple of places where *SpinBoxSetting was called with a boolean
instead of an integer.  The spinbox code was changed in 2017, and the
change expected this implicit conversion to occur for backward
compatibility.  Convert callers so the backward compatibility is no
longer necessary.

https://clang.llvm.org/extra/clang-tidy/checks/readability-implicit-bool-conversion.html