Read signal strength, bit error rate, etc with the DVBv5 API.
Read signal strength, signal to noise ratio, bit error rate and
the uncorrected block count now first with the DVBv5 API calls;
if that fails then try with the DVBv3 API calls.
Removed conditional compilation based on the DVB API version.
Previously this was done first with the DVBv3 and then with
the DVBv5 calls if the DVBv5 code was compiled in.

Refs #12638