tidy: Use the correct type for loop counters.
The clang-tidy "loop variable too small" check pointed out many places
where the variable used in a for loop is too small for upper limit on
the value being counted.  Almost all of these are for using a uint to
count up to x.size().  Since the size() function returns a value of
type size_t, the trivial solution to these is to change the counter
type to size_t so its always guaranteed to be the right size.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-too-small-loop-variable.html