tidy: Rename zmliveplayer showMenu function and mark it as an override.
The clang-tidy "virtual near miss" check pointed out one function that
appears should have been an override but didn't match exactly.  The
first letter of this function is lowercase with the first letter of
the function in the parent class is uppercase.  They both take the
same arguments.  Change the capitalization on this function name and
mark it as an override.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-virtual-near-miss.html