Fix UI File Browser to show sizes in tenths.
The clang-tidy "integer division result used by floating point" test
indicated that MythUIFileBrowser::FormatSize was taking the size of a
file, performing an integer division, converting the result to a
double, and printing it.  This is guaranteed to always produce an
integer result, so the cast to double is superfluous.  The intended
result seems to be to print out some number of decimal places and not
just an integer, so cast the file size to a double before dividing the
number.  Also tweak the parameters to QString::arg so that it will
only print out a single place after the decimal point.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-integer-division.html