tidy: Eliminate unnecessary copying of objects. (libs)
The clang-tidy "unnecessary value param" check pointed out many places
where objects were passed by copying when they could be passed by
const reference.  The vast majority of these fixes convert a QString
object into a const QString& reference.  There are also a number of
other objects that get passed by reference instead of copying.

This check also added '::std::move' to a number of object uses.  If
the last reference to an object is to assign that object to a variable
or to use it as a parameter in a function call, then its a waste of
cycles to duplicate the object for the operation.  Just pass ownership
of the already existing object to the target of the operation.

Changes made by the clang-tidy program.

https://clang.llvm.org/extra/clang-tidy/checks/performance-unnecessary-value-param.html