Change DVB-T to DVB-T2 and DVB-S to DVB-S2 if delivery system not defined.
If the delivery system is not defined in database table capturecard then mythbackend uses the tuner in the state it is in.
If the current delivery system is DVB-T and the tuner is capable of DVB-T2 then the delivery system is changed to DVB-T2.
This allows reception of both DVB-T and DVB-T2 multiplexes.
Same for DVB-S and DVB-S2.
This allows a mythbackend of the master to run with a v30 database, where the value of
capture card column inputname is either empty or "DVBInput".
There is a message in the console log when this happens.

Refs #13014