Starting with Qt5.10, replace calls to rand/srand with QRandomGenerator.
The rand/srand calls are considered obsolete.  Qt5.10 introduced
QRandomGenerator as a replacement, so take advantage of that.  Qt
provides a global random generator that was securely seeded from the
system random entropy pool.  This seeding function is expensive, so
calls to reseed the random number generator at thread creation or when
gallery views are created have been removed.