Change scheduler warning message about Listings source to Video source
The scheduler gives, at the start of mythbackend, a warning message
for each "Listings source" that is not connected to a capturecard.
The message is relevant but confusing because in mythtv-setup the
card input can be connected to a video source and this warning message
clearly refers to the video source.