Only real tuners for Input device of Channel Scan.
In mythtv-setup / Channel Editor / Channel Scan the input device
(capture card) can be selected from a list of input devices
that are connected to the video source.
This list did also contain the multirec slave tuner instances which
is not relevant for mythtv-setup and can make for very long lists.