Do not overwrite modulation system in multiplex with value from tuner.
A DVB-S multiplex can be received by a tuner configured for a DVB-S2 delivery system.
The removed line of code did overwrite the modulation system in
the multiplex with the delivery system currently configured in the tuner.
This caused the modulation system in all DVB-S/S2 multiplexes to be set to DVB-S2.

Refs #13014