Enable Schedule as Group by default.
This also includes the case when the maximum recordings is set to 1.
Both the core feature scheduling feature and the on-demand addition of
inputs feature are stable and and should be enabled by default for all
users.

(cherry picked from commit 895daa592802ea3b16d8f1e583d69cc8fea5f2c3)