Fix a handful of new gcc9 "implicitly-declared <blah> is deprecated" warnings.
These warnings now occur when an object has a custom destructor but
not a copy constructor or a custom operator= function.  The thought is
that the custom destructor was probably written to free memory managed
by the object, and that the default copy constructor and operator=
function will simply copy the pointer value from one object to the
other.  That is a setup for an invalid memory reference to occur after
the first of the two objects is destroyed.

One of the functions fixed here need to copy memory as part of its
copy constructor and operator= functions.  There's already a copy
function available on the object, so use that.  The rest of the
functions fixed by this patch can simply declare that the default copy
constructor is sufficient.