Select best default delivery system when creating new capture card.
Previously, when creating a new capture card in mythtv-setup
the default delivery system was the first one of the available
delivery systems for that card. Unfortunately, if a device is
capable of e.g. both DVB-T and DVB-T2 then DVB-T is first on the
list and thus the default even if DVB-T2 is usually the better choice.
Idem for DVB-S and DVB-S2.
With this fix the better choice is now the default.
Note that when editing an existing capture card the value
for the delivery system in the database is always the default,
even if it is DVB-T or DVB-S.

Refs #13014