Remove unused HasLock call on tuner that is not yet tuned
In DVBSignalMonitor::DVBSignalMonitor function HasLock is called.
At the time the function is called the tuner has not been
tuned yet so checking for the lock status is not useful.
There is nothing done with the results of this call except
giving an error message if the call fails.
The reason to remove this call is that the frontend
"Silicon Labs Si2168", used in the MyGica T230, returns
the EREMOTEIO error on FE_READ_STATUS if the tuner is not tuned.

Refs #13447