Ask only once for non-conflicting channel number
In mythtv-setup, when there is a channel number conflict after a channel scan, it is
posssible to select "Insert manually" and then enter a non-conflicting channel number yourself.
However, when this is done another popup dialog appears to let you do this once again.
This fix skips the second dialog when there is already a non-conflicting channel number chosen.