Use transport id as service id in MPTS channels
In mythtv-setup, it is possible to create MPTS channels that
can record the complete transport stream without any pid filtering.
In the channel list, the service ID is used as channel number when
there is no logical channel number information present.
For the MPTS channels, who do not have a service id,  this means
that the channel number must always be entered manually, both on
an initial scan and on a rescan.
This fix uses the transport stream ID as the service ID; the default
channel number of the MPTS channels is then the transport stream ID.

N.B. It is possible that the transport stream id does conflict
with either a logical channel number or a service id but when that
happens the channel number conflict resultion dialogs are still there.