Set status of MPTS recording
Set the status of an MPTS full-transport stream recording to Recording
also in the current recording, tv_rec.m_curRecording.
If this is not done the status of the current recording remains Pending.
This does not affect the recording itself but it causes the recording
to be flagged as Failed when the recording is finished, even
when the recording itself is perfect.