MythNVDECContext: Fix NVDEC deinterlacing of H.264 material
- the H.264 parser flagged stream changes for the raw, underlying stream
rate (before any double rate deinterlacing in the decoder). This frame
rate triggered MythPlayer to lock the scan as it is in the normal range
for interlaced material.
- MythPlayer is now notified of the adjusted framerate - although this
may need tweaking as the user now doesn't know whether the stream is
actually interlaced or progressive...
- N.B. further work is required for single rate decoder deinterlacing
which will still be locked as interlaced and hence actual frame values
will be ignored.