Improved interlaced detection
- the basic premise here is that we don't want to setup deinterlacers
unnecessarily (to save resources and avoid confusion in the logs etc)
and that FFmpeg is pretty good these days at getting interlaced flags
correct

- so don't use detection based on size/frame rate on stream changes, set
the initial state as interlaced but with the scan tracker as progressive
(if that makes sense). The scan is immediately corrected on detection of
a progressive frame and no deinterlacers are created. We set to
interlaced only to account for those cases where the frame rate is
detected incorrectly on startup (usually 29.97 as 59.98) and we
initially think the file is progressive.
- testing with my 'test suite' of curious files and formats, this gets it
correct for all but 1 of about 120 files - which is better than NVDECs
detection. The one misdetected file is clearly an interlaced NTSC stream
that is not flagged as such - all playes get it wrong. NVDEC doubles the
frame rate but doesn't deinterlace it.