VAAPI: Fix 10bit copyback display
- we now create our own VAAPI frames context so that we can set the SW
format to something that FFmpeg won't object to.
- this at least makes the decoder setup comparable for direct rendering
and decode only.

N.B. On my Coffee Lake test setup with Debian Buster and libva 2.4.0 -
this requires the i965-va-driver-shaders package (non-free) instead of
the standard i965-va-driver package. Otherwise libva returns a 'not
implemented' error to FFmpeg for the vaGetImage call.