UPnP Serializer: Eliminate false 304 response statuses
Affects Services API responses.

For 'simple' Name=Value responses (Bool/Int/String/etc.),
include Value in the hash calculated for the ETag.

Without this, the ETag never changed and the backend always
sent 304/Not Modified on subsequent queries (assuming the
client returned the "If-None-Match" header.) For example:

  http://hostname:6544/Myth/GetSetting?Key=UserJobDesc4&HostName=_GLOBAL_

the backend always returned:

  'ETag': '"da39a3ee5e6b4b0d3255bfef95601890afd80709"'

regardless of the value in UserJobDesc4.