VAAPI: Fix scaling of 1088 video when using copyback
- round 387 of the VAAPI merry-go-round
- 1088 video was displaying the garbage line at the bottom of the screen
when using copy-back
- we need to create the VAAPI frames context with the coded_height of
the source material so that full 1088 frames are returned - with the
VideoOutput class just ignoring the garbage. If 1080 is used, the driver
appears to scale the 1088 frame into a 1080 frame when copying back to
main memory.
- requires a reversion back to initialising the decoder in GetFormat as
the coded_height is not available/ready in HwDecoderInit
- re-introduces the problem of multiple calls to InitialiseContext2 when
using copyback for some VP8 files - but this appears to be harmless, if
annoying.
- using the coded_height also fixes the previous problems seen with some
files when InitialiseContext2 was called from GetFormat2