Set MPTS full transport stream channel to unencrypted
Previously, the value  of "m_is_encrypted" value was copied from the
first channel in the transport stream and that channel could be encrypted.
This fix insures that the "Add full Transport Stream channels" option
in mythtv-setup/Channel Scan will always add the MPTS channel even if
the "Unencrypted Only" option is selected.