VideoBuffers: Optimise number of buffers for reference frames in use
- parse the number of reference frames from the AVCodecContext when
initialising the video stream and pass the value through to the
VideoOutput classes via the player.
- use the H.264 parser to check the number of ref frames (but don't use
it to trigger a stream change)
- use the number of ref frames to determine the number of video buffers.
Only currently impacts VAAPI and software decode.
- HEVC, VP8 and VP9 default to their maximum values (i.e. they are not
parsed - not even sure they can be)
- otherwise the default is 2

This significantly reduces the number of video buffers for a lot of
content - and hence reduces memory consumption.