Add 29700000 to list of DVB-S/S2 symbol rates
The DVB-S/S2 symbol rate popup list is defined in two places;
in transporteditor.cpp for editing transports, and in
channelscanmiscsettings.h for the "Full Scan" parameters.
The value of 29700000 is used for Swiss TV on Hotbird 13E.
This value is now added in both lists.
Added 22500000 to the list in the transport editor,
to make it identical to the list in the "Full Scan".