Name of MPTS service created with tsid from SDT
The MPTS full-transport stream service uses the transport stream id
as part of the name. This is very convenient if you have more than
one MPTS service in the channel list.
Note that the MPTS service is not found in the scan but is created
on request for each transport stream as a copy of the first service
found in the scan of that transport stream.
It can happen that the first service is not defined in the PAT/PMT but
only in the SDT. In that case the old code, which used the transport
stream id as found in the PAT, fails and the MPTS service is then
called MPTS_UNKNOWN.
This fix uses the transport stream id from from the SDT; if the first
service is not in the SDT, which can also happen, the transport
stream id from the PAT is used.