New mythtv-setup option: Complete scan data required
This is a new option in the mythtv-setup Channel Editor/Channel Scan page.
If set, only services that are present in the PAT, the PMT
and the SDT and that have a name will be added during the scan.
This option is very useful when doing scans on satellites.
On satellites there are typically many incomplete channels,
such as feeds and offline channels, that are not useful in a MythTV system.
The result of this using this option is that you get the same set of
channels in MythTV as what is listed on lyngsat.com or kingofsat.com.