VideoFrame: Default to the reported video height
- rather than always forcing a 16pixel height alignment, use the
reported height but allocate the frame buffers with the extra alignment
to minimise any subsequent re-allocations and ensure FFmeg does not
overrun
- make VideoBuffers::ReinitBuffer static to avoid a round trip of calls
from the decoder. There should be no need for locking as the frame is
held by the decoder.
- the only complication is that FFmpeg seems to require 1088 type
heights for some (most) codecs - so adjust the frame allocation as
necessary

- fixes a few unnecessary scaling rendering issues - notably 1080 HEVC
software decode