Use ca0 when frontend1 is in use but no ca1 is available.
Multi-standard cards such as TBS9528 have two frontends,
frontend0 for DVB-T/T2/C and frontend1 for DVB-S/S2/DSS.
There is only one demux etc: demux0, dvr0 and ca0.
When frontend1 is selected then demux0/dvr0/ca0 should
be used with that frontend.
This has been fixed in ticket #13052 for demux0 and dvr0
but not yet for ca0.
This fix selects ca0 when ca<x> for frontend<x> is not
available, similar to what is done for demux and dvr.

Refs #13052