Add sort mode "Multiplex Frequency" in mythtv-setup Channel Editor.
In addition to being able to sort on Channel Name or Channel Number
it is now possible to sort the channels in the Channel Editor
of mythtv-setup on Multiplex Frequency, with the service id
as the secondary sort key.
This makes it much easier to compare the channel scan results of
mythtv-setup with the channnels as listed by kingofsat.net and
lyngsat.com as this is the sorting order these websites use.