Clear EIT channel locks on master startup
EIT channels can be left locked when mythbackend is
terminated with a SIGKILL.
This means that there will be no new EIT information
for these channels when mythbackend is started again.
The only way to get the channels unlocked at short
notice is to execute the appropriate sql commands manually.
The function to clear the channel locks was already
implemented, but it was never called.
It is now called on startup of the master backend.