Channel list sorted on frequency in mythtv-setup output
The list of channels written to standard output by mythtv-setup
after a succesful scan is now sorted in order of increasing freqency.
This makes it easier to compare with channel lists from previous scans
and, for satellites, it also makes it easier to compare the output of
mythtv-setup with the channel information on lyngsat.com and kingofsat.net.