Function value of ShowManualChannelPopup
The ShowManualChannelPopup dialog of mythtv-setup
has the choices OK, Edit, Cancel and Cancel All.
This dialog is used when there is a conflict in updating
channels after a channel scan.
With this fix the dialog returns kOctCancel for the Cancel
button and kOctCancelAll for the Cancel All button.
Previously, kOctCancelAll was returned for both the Cancel
and the Cancel All button.