Improved dialog text for new channels
In mythtv-setup, when new channels are found there is a popup message.
This message is now "Found %n new %1 channel(s).".
Previously there was also "non-conflicting" mentioned.
This was not very clear as at this stage in the scanning process
the only thing that is checked is that the service id's of the
new channels are unique within the video source; the channel
numbers are not yet known.
In mythtv-setup, the terms "conflicting" and "non-conflicting"
generally refer to channel numbers.