Sorted channel lists
The terminal output of mythtv-setup has channel lists that
are now all sorted in order of increasing frequency.
In addition to the complete list of channels there are now
also separate lists of:
- off-air channels
- updated old channels
- skipped old channels
- inserted new channels
- skipped new channels
The lists with updated and inserted channels
also have the channel number.
There are no intentional functional changes.

Refs #13472