Fix MythSytemEvent class of MythTV's python bindings
MythSytemEvent:

An event is identified by the regex [A-Z0-9_]* .
This includes the system event `KEY_01`.

BEEventConnection:

The class `BEEventConnection` has optional arguements like
`timeout` an `level`. Name then if the arguement order is not followed strictly.

Typo in the call of
"SystemEvent(event['event'], inst.db).command(event)"

Note:
'inst.db' is not a valid class at this level

Uppercase of system event scripts:

According wiki, the default substitution of parameters is given by uppercase,
like '%STARTTIMEUTC%' or %SENDER%'. Stick to that convention.