Various fixes of MythTV Python Bindings in mythproto.py.
Commit eebe25c introduced the new setting 'BackendServerAddr' instead
of 'BackendServerIP[6]'. See ticket #13082.
Fix another occurence of  'BackendServerIP[6]' in class BECache.

In mythtv's protocol #84 the event 'UPDATE_FILE_SIZE' changed to use
'recordedid' instead of tuple ('chanid', 'starttime').
This commit adds an optional parameter 'recordedid' to ftopen(),
but still allows the optional ('chanid', 'starttime') tuple, and
fixes the handling of this event.
See ticket #12365, comment:5.

Fix the optional arguements when openening the control socket in the
class 'FileTransfer'. We need to enable receive events, if we want
listen to them. Long time ago, the commit cd23715 changed this
behaviour of the BEEvent class.

The method 'FileOps.downloadTo()' listen only for events
'DOWNLOAD_FILE UPDATE'. In short file transfers, the event
'DOWNLOAD_FILE FINISHED' is sent without an update event.
Listen for both events when allocating an eventlock.